Add internal checks for inputs
We can't throw errors like we normally do because errors are also captured by gradethis and turned into feedback messages. So when we're checking that authors have provided correctly formatted arguments, we need to capture the errors and handle them ourselves. `assert_internally()` does that by printing the error as a message (should appear in logs) and returning an "interal_feedback_error" with a friendly user-facing message.